backend-marlbruh/requirements.txt
Django
djangorestframework
django-cors-headers
Pillow
gunicorn
tzdata
ebooklib